Move Iceberg procedures to separate class
